pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "prometheus-client>=0.21.1",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "python-dotenv>=1.1.0",
    "pytz>=2025.2",
    "redis>=6.0.0",
    "websockets>=12.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.8",
]

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
